+++ Sources/BezierPathExtensions.m
@@ -72,7 +72,7 @@ static double sRanges[8] = {0, 1, 0, 1, 0, 1, 0, 1};
     float alpha = inDegrees / 180.0 * M_PI;
     float dx = cos(alpha);
     float dy = sin(alpha);
-    int i, n = [self elementCount];
+	NSUInteger i, n = [self elementCount];
     NSPoint points[3];
     float d, dmin = 0, dmax = 1;
     BOOL first = YES;
